% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fucnction_HW.R
\name{adelaide_offence_level_3}
\alias{adelaide_offence_level_3}
\title{Correlation Finder
Finds the correlation between the offence count in two suburbs
\code{adelaide_Offence_level_3} The functions visualizes the correlation in offence count between the two input suburbs for a level 3 offence.}
\usage{
adelaide_offence_level_3(crime_data, offence_description, suburbs)
}
\arguments{
\item{crime_data}{A data.table object with the following columns:
"date" (POSIXct), "suburb" (chr), "postcode" (chr), "offence_level_1" (chr),
"offence_level_2" (chr), "offence_level_3" (chr), "offence_count" (num).}

\item{offence_description}{A character string of type offence_level_3.}

\item{suburbs}{A two-element character vector. Each element is the name (UPPERCASE)
of an SA suburb.}
}
\value{
A ggplot object showing the correlation in offence count between the two input suburbs.
}
\description{
Correlation Finder
Finds the correlation between the offence count in two suburbs
\code{adelaide_Offence_level_3} The functions visualizes the correlation in offence count between the two input suburbs for a level 3 offence.
}
\examples{
adelaide_offence_level_3(crime_2012,"OFFENCES AGAINST PROPERTY", vec)
}
