% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{create_patient_record}
\alias{create_patient_record}
\title{Create patient record}
\usage{
create_patient_record(full_name, weight_lb, height_in)
}
\arguments{
\item{full_name}{A character string containing two names seperated by a space}

\item{weight_lb}{A number. The patient's weight in pounds}

\item{height_in}{A number. The patients height in inches.}
}
\value{
A list of first_name, last_name, weight_kg and height_m
}
\description{
\code{create_patient_record} takes a person’s full name (FirstName LastName), weight
  in lb and height in inches and converts it into a list with the person’s first
  name, person’s last name, weight in kg and height in m. There are 2.2 lb in a kg.
  There are 39.4 inches in a metre.
}
